Projects/3. qa_bot_pro/requirements.txt
streamlit>=1.38.0
langchain>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
google-generativeai>=0.8.0
docx2txt>=0.1
pypdf>=5.0.0
python-dotenv>=1.0.0